+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.3.0
+
+Released on 2024-09-17.
+
+- New command `remove` to remove an entry from the library and its note and PDF files.
+
 ## v0.2.1
 
 Released on 2024-09-17.